@@ -1,11 +1,12 @@
 import './index.css';
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { App } from './App';
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!); 
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root'),
+  </React.StrictMode>
 );
