@@ -28,19 +28,19 @@ export class QRCode extends React.Component<Props> {
           '&:hover': {
             transform: 'scale(1.02)',
             boxShadow: '0 8px 24px rgba(0,0,0,0.12)',
-          }
+          },
         }}
       >
-          <img 
-            alt="WireGuard QR code" 
-            src={this.uri.current}
-            style={{
-              display: 'block',
-              width: '250px',
-              height: '250px'
-            }}
-          />
-        </Box>
+        <img
+          alt="WireGuard QR code"
+          src={this.uri.current}
+          style={{
+            display: 'block',
+            width: '250px',
+            height: '250px',
+          }}
+        />
+      </Box>
     );
   }
 }
